+++ app/library/System/Cron/ScheduleTime.php
@@ -2,21 +2,29 @@
 
 namespace System\Cron;
 
-use Model\CronLog\CronLog;
-
 class ScheduleTime
 {
   private $call_back;
   private $params = [];
   private $time;
   private $event_name = 'animus';
-
+  private $time_exect;
+  private $time_name = '';
+  private $animusly = false;
 
   public function __construct($call_back, $params, $time)
   {
     $this->call_back = $call_back;
     $this->params = $params;
     $this->time = $time;
+    $this->time_exect = [
+      [
+        'D' => date('D', $this->time),
+        'd' => date('d', $this->time),
+        'h' => date('H', $this->time),
+        'm' => date('i', $this->time)
+      ]
+    ];
   }
 
   public function eventName($val)
@@ -25,13 +33,37 @@ class ScheduleTime
     return $this;
   }
 
-  private function exect(array $events)
+  public function animusly($run_as_animusly = true)
   {
+    $this->animusly = $run_as_animusly;
+    return $this;
+  }
+
+  public function isAnimusly(): bool
+  {
+    return $this->animusly;
+  }
+
+  public function getEventname()
+  {
+    return $this->event_name;
+  }
+
+  public function getTimeName()
+  {
+    return $this->time_name;
+  }
+
+  // TODO: get next due time
+
+  public function exect()
+  {
+    $events = $this->time_exect;
+
     $dayLetter  = date("D", $this->time);
     $day        = date("d", $this->time);
     $hour       = date("H", $this->time);
     $minute     = date("i", $this->time);
-
 
     foreach ($events as $event) {
       $eventDayLetter = $event['D'] ?? $dayLetter; // default day letter every event
@@ -44,7 +76,6 @@ class ScheduleTime
         // stopwatch
         $watch_start = microtime(true);
 
-        // TODO: more call back support
         $out_put = call_user_func($this->call_back, $this->params) ?? [];
 
         // stopwatch
@@ -58,18 +89,21 @@ class ScheduleTime
 
   public function justInTime()
   {
-    return $this->exect([
+    $this->time_name = __FUNCTION__;
+    $this->time_exect = [
       [
         'D' => date('D', $this->time),
         'd' => date('d', $this->time),
         'h' => date('H', $this->time),
         'm' => date('i', $this->time)
       ]
-    ]);
+    ];
+    return $this;
   }
 
   public function everyTenMinute()
   {
+    $this->time_name = __FUNCTION__;
     $minute = [];
     for ($i=0; $i < 60; $i++) {
 
@@ -82,12 +116,15 @@ class ScheduleTime
       }
     }
 
-    return $this->exect($minute);
+    $this->time_exect = $minute;
+
+    return $this;
   }
 
   public function everyThirtyMinutes()
   {
-    return $this->exect([
+    $this->time_name = __FUNCTION__;
+    $this->time_exect = [
       [
         'd' => date('d', $this->time),
         'h' => date('H', $this->time),
@@ -98,11 +135,14 @@ class ScheduleTime
         'h' => date('H', $this->time),
         'm' => 30,
       ],
-    ]);
+    ];
+
+    return $this;
   }
 
   public function everyTwoHour()
   {
+    $this->time_name = __FUNCTION__;
 
     $thisDay = date("d", $this->time);
     $hourly = []; // from 00.00 to 23.00 (12 time)
@@ -116,12 +156,14 @@ class ScheduleTime
       }
     }
 
-    return $this->exect($hourly);
+    $this->time_exect = $hourly;
+    return $this;
   }
 
   public function everyTwelveHour()
   {
-    return $this->exect([
+    $this->time_name = __FUNCTION__;
+    $this->time_exect = [
       [
         'd' => date('d', $this->time),
         'h' => 0,
@@ -132,11 +174,13 @@ class ScheduleTime
         'h' => 12,
         'm' => 0,
       ]
-    ]);
+    ];
+    return $this;
   }
 
   public function hourly()
   {
+    $this->time_name = __FUNCTION__;
     $hourly = []; // from 00.00 to 23.00 (24 time)
     for ($i=0; $i < 24; $i++) {
       $hourly[] = [
@@ -146,60 +190,75 @@ class ScheduleTime
       ];
     }
 
-    return $this->exect($hourly);
+   $this->time_exect = $hourly;
+   return $this;
   }
 
   public function hourlyAt(int $hour24)
   {
-    return $this->exect([
+    $this->time_name = __FUNCTION__;
+    $this->time_exect = [
       [
         'd' => date('d', $this->time),
         'h' => $hour24,
         'm' => 0
       ]
-    ]);
+    ];
+
+    return $this;
  }
 
   public function daily()
   {
-    return $this->exect([
+    $this->time_name = __FUNCTION__;
+   $this->time_exect = [
       // from day 1 to 31 (31 time)
       ['d' => date('d'), 'h' => 0, 'm' => 0],
-    ]);
+    ];
+
+    return $this;
   }
 
   public function dailyAt(int $day)
   {
-    return $this->exect([
+    $this->time_name = __FUNCTION__;
+    $this-> time_exect = [
       [
         'd' => $day,
         'h' => 0,
         'm' => 0,
       ]
-    ]);
+    ];
+
+    return $this;
   }
 
   public function weekly()
   {
-    return $this->exect([
+    $this->time_name = __FUNCTION__;
+   $this->time_exect = [
       [
         'D' => "Sun",
         'd' => date('d', $this->time),
         'h' => 0,
         'm' => 0
       ],
-    ]);
+    ];
+
+    return $this;
   }
 
   public function mountly()
   {
-    return $this->exect([
+    $this->time_name = __FUNCTION__;
+    $this->time_exect = [
       [
         'd' => 1,
         'h' => 0,
         'm' => 0
       ],
-    ]);
-  }
+    ];
 
+    return $this;
+  }
 }