+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Global function
+- Better chain code in cron shedule
+- Add command cron:list to show all cron register
 
 ### Changed
 - Help command now create by every single command child class
 - Change file from MakerCommand.php to MakeCommand.php
+- Rename file from Schadule.php to Schedule.php
 
 ## [0.1.5] - 2021-05-21
 ### Added